refactor(logger): modularize, polish console/evlog output, fix error logging
- split index.ts into core/plugin/console/pino/evlog/utils modules
- rename createPinoLoggerAdapter -> createPinoAdapter (symmetry with createEvlogAdapter)
- console adapter: 256-color palette, aligned multi-line fields, error stack rendering; serialize Error fields in production JSON (was {})
- evlog adapter: derive tag/service bracket, drop spurious tag field, emit wide events via object form (no bogus 'in 0ms' duration)
- emit a single wide event per interaction (remove duplicate immediate error logs)
- add useRootLogger() for logging outside an interaction scope
- capture non-Error structured args from Seyfert internal logs as details; always intercept (remove interceptInternal option)
- rewrite README around the wide-event model
