+++ packages/logger/src/utils.ts
@@ -1,0 +1,32 @@
+export function asRecord(value: unknown): Record<string, unknown> {
+	return value && typeof value === 'object' ? (value as Record<string, unknown>) : {};
+}
+
+export function isLogData(value: unknown): value is Record<string, unknown> {
+	return (
+		!!value &&
+		typeof value === 'object' &&
+		!(value instanceof Date) &&
+		!(value instanceof Error) &&
+		!Array.isArray(value)
+	);
+}
+
+export function getString(value: unknown): string | undefined {
+	return typeof value === 'string' && value.length > 0 ? value : undefined;
+}
+
+export function getNumber(value: unknown): number | undefined {
+	return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
+}
+
+export function getStringField(value: unknown, field: string): string | undefined {
+	return getString(asRecord(value)[field]);
+}
+
+export function stripUndefined<T extends Record<string, unknown>>(value: T): T {
+	for (const key of Object.keys(value)) {
+		if (value[key] === undefined) delete value[key];
+	}
+	return value;
+}
